+++ packages/bridge-controller/src/index.ts
@@ -40,6 +40,7 @@ export {
 export type {
   ChainConfiguration,
   L1GasFees,
+  NonEvmBalanceError,
   NonEvmFees,
   GasMultiplierByChainId,
   FeatureFlagResponse,